docker: update to 17.06 with mobynit support
From now on we'll keep all the development and patches of the docker daemon in
its own repo (resin-os/docker at the time of writing). All the patches that
existed in this recipe were applied to our tree and also the version was
updated to 17.06.

Furthermore, this version of moby comes with mobynit, a new command that can be
used as PID 1. This allows one to create hostapp enabled systems by booting
into a container. Since at the time mobynit runs there is no userspace set up
it gets statically compiled and UPXed to save space.

The resulting binary is deployed in the special /bootstrap directory. Any files
that are in that directory will be copied out of the docker containers in the
host. This directory should be used for files that need to be read by the
bootloader or other stages that can't boostrap docker yet.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>