docker: set bridge subnet to 10.114.101.1/24
We change the default subnet of docker to avoid conflicts with the local
network which will cause one of the two conflicting networks to become
unroutable. Docker attempts to find an unused subnet on startup but this is not
enough to avoid coflicts in the case of resin because the network might not
have come up when docker starts (e.g slow wifi/DHCP) or might change after it
has come up (e.g resin wifi connect). Furthermore, even if the subnet
automatically selected by docker is not directly accessible from the host, it
might be through the default gateway. So no matter what we do there is a
possibility to shadow a useful subnet.

This change attempts to minimise the chances of this happening by reducing the
size of the subnet to /24 instead of /16 and also moving it to a fairly
uncommon network, 10.114.101.0. This will solve most of the networking issues
we see today. However, it should become configurable in the future.

On top of that, we split the subnet between the host and application docker
daemon. The application docker daemon allocates containers in the lower
10.114.101.0/25 subnet and the host container allocates containers in the upper
10.114.101.128/25 subnet.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>